+++ frontend/src/modules/constructor/ContentPizza.vue
@@ -1,0 +1,28 @@
+<template>
+    <div class="content__pizza">
+        <label class="input">
+            <span class="visually-hidden">Название пиццы</span>
+            <input type="text" name="pizza_name" placeholder="Введите название пиццы">
+        </label>
+
+        <div class="content__constructor">
+            <div class="pizza pizza--foundation--big-tomato">
+                <div class="pizza__wrapper">
+                    <div class="pizza__filling pizza__filling--ananas"></div>
+                    <div class="pizza__filling pizza__filling--bacon"></div>
+                    <div class="pizza__filling pizza__filling--cheddar"></div>
+                </div>
+            </div>
+        </div>
+
+        <div class="content__result">
+            <p>Итого: 0 ₽</p>
+            <button type="button" class="button" disabled>Готовьте!</button>
+        </div>
+    </div>
+</template>
+
+<style lang="scss" scoped>
+    @import '@/assets/scss/fonts.scss';
+    @import '@/assets/scss/style.scss';
+</style>
